adding string compression class
