+++ src/Program.cs
@@ -1,12 +1,11 @@
 ﻿using System;
-using System.IO;
-using System.IO.Compression;
 using NATS.Client;
 using System.Text;
 using NLog;
 using NLog.Config;
 using openrmf_msg_checklist.Models;
 using openrmf_msg_checklist.Data;
+using openrmf_msg_checklist.Classes;
 using MongoDB.Bson;
 using Newtonsoft.Json;
 
@@ -47,7 +46,7 @@ namespace openrmf_msg_checklist
                     string msg = JsonConvert.SerializeObject(art).Replace("\\t","");
                     // publish back out on the reply line to the calling publisher
                     logger.Info("Sending back compressed Checklist Data");
-                    c.Publish(natsargs.Message.Reply, Encoding.UTF8.GetBytes(CompressString(msg)));
+                    c.Publish(natsargs.Message.Reply, Encoding.UTF8.GetBytes(Compression.CompressString(msg)));
                     c.Flush(); // flush the line
                 }
                 catch (Exception ex) {
@@ -61,30 +60,6 @@ namespace openrmf_msg_checklist
             // arriving immediately.
             logger.Info("setting up the openRMF checklist subscription");
             IAsyncSubscription asyncNew = c.SubscribeAsync("openrmf.checklist.read", readChecklist);
-        }
-        /// <summary>
-        /// Compresses the string.
-        /// </summary>
-        /// <param name="text">The text.</param>
-        /// <returns></returns>
-        public static string CompressString(string text)
-        {
-            byte[] buffer = Encoding.UTF8.GetBytes(text);
-            var memoryStream = new MemoryStream();
-            using (var gZipStream = new GZipStream(memoryStream, CompressionMode.Compress, true))
-            {
-                gZipStream.Write(buffer, 0, buffer.Length);
-            }
-
-            memoryStream.Position = 0;
-
-            var compressedData = new byte[memoryStream.Length];
-            memoryStream.Read(compressedData, 0, compressedData.Length);
-
-            var gZipBuffer = new byte[compressedData.Length + 4];
-            Buffer.BlockCopy(compressedData, 0, gZipBuffer, 4, compressedData.Length);
-            Buffer.BlockCopy(BitConverter.GetBytes(buffer.Length), 0, gZipBuffer, 0, 4);
-            return Convert.ToBase64String(gZipBuffer);
         }
 
         private static ObjectId GetInternalId(string id)
